+++ src/steps/domains/converters.ts
@@ -20,7 +20,7 @@ export function createDomainEntity(data: admin_directory_v1.Schema$Domains) {
         id: domainName,
         displayName: domainName,
         name: domainName,
-        domainName,
+        emailDomain: domainName,
         primary: data.isPrimary === true,
         verified: data.verified === true,
       },